refactor: rename ControlPanelLayout to ControlPanel
