--- src/components/landing/CaseStudy/ControlPanelLayout/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ControlPanelLayout";
-export * from "./ControlPanelLayout";
